+++ database/migrations/administrator/2017_06_01_183014_create_equipment_table.php
@@ -15,8 +15,8 @@ class CreateEquipmentTable extends Migration
     {
         Schema::connection('administrator')->create('equipment', function (Blueprint $table) {
 
-            $table->increments('id');
-            $table->unsignedInteger('equipment_category_id');
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('equipment_category_id');
             $table->string('name')->unique();
             $table->string('description',500)->nullable();
