+++ database/migrations/administrator/2018_05_30_200339_create_level_alert.php
@@ -15,8 +15,8 @@ class CreateLevelAlert extends Migration
     {
        Schema::connection('administrator')->create('level_alert', function (Blueprint $table) {
 
-            $table->increments('id');
-            $table->unsignedInteger('alert_id');
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('alert_id');
 
             $table->string('name')->nullable();
             $table->string('code')->unique();
