+++ database/migrations/config/2017_02_06_210815_create_var_for_station_config.php
@@ -15,9 +15,9 @@ class CreateVarForStationConfig extends Migration
     {
       Schema::connection('config')->create('var_for_station', function (Blueprint $table) {
 
-          $table->increments('id');
-          $table->integer('station_id');
-          $table->integer('variable_id');
+          $table->bigIncrements('id');
+          $table->unsignedBigInteger('station_id');
+          $table->unsignedBigInteger('variable_id');
           $table->float('minimum', 8, 4)->nullable();
           $table->float('maximum', 8, 4)->nullable();
           $table->float('previous_difference', 8, 4)->nullable();
