+++ database/migrations/temporaryWork/2017_04_10_222403_create_exist_fact_table.php
@@ -15,9 +15,9 @@ class CreateExistFactTable extends Migration
     {
         Schema::connection('temporary_work')->create('exist_fact_weather', function (Blueprint $table) {
 
-            $table->integer('station_sk')->nullable();
-            $table->integer('date_sk')->nullable();
-            $table->integer('time_sk')->nullable();
+            $table->unsignedBigInteger('station_sk')->nullable();
+            $table->unsignedBigInteger('date_sk')->nullable();
+            $table->unsignedBigInteger('time_sk')->nullable();
 
             $table->string('date_time')->nullable();
